Make 'serve' the default command in docker image
