+++ mlflow/models/docker_utils.py
@@ -5,8 +5,6 @@ import logging
 import mlflow
 import mlflow.version
 from mlflow import pyfunc, mleap
-from mlflow.exceptions import MlflowException
-from mlflow.protos.databricks_pb2 import RESOURCE_DOES_NOT_EXIST, INVALID_PARAMETER_VALUE
 from mlflow.utils.file_utils import TempDir, _copy_project
 from mlflow.utils.logging_utils import eprint
 
@@ -47,66 +45,13 @@ ENV JAVA_HOME=/usr/lib/jvm/java-8-openjdk-amd64
 RUN conda install gunicorn;\
     conda install gevent;\
 
-{install_mlflow}
-{custom_setup_steps}
-
 # Set up the program in the image
 WORKDIR /opt/mlflow
 
-# start mlflow scoring
-ENTRYPOINT ["python", "-c", "import sys; from mlflow.sagemaker import container as C; \
-C._init(sys.argv[1])"]
+{install_mlflow}
+{custom_setup_steps}
+{entrypoint}
 """
-
-
-def _get_preferred_deployment_flavor(model_config):
-    """
-    Obtains the flavor that MLflow would prefer to use when deploying the model.
-    If the model does not contain any supported flavors for deployment, an exception
-    will be thrown.
-
-    :param model_config: An MLflow model object
-    :return: The name of the preferred deployment flavor for the specified model
-    """
-    if mleap.FLAVOR_NAME in model_config.flavors:
-        return mleap.FLAVOR_NAME
-    elif pyfunc.FLAVOR_NAME in model_config.flavors:
-        return pyfunc.FLAVOR_NAME
-    else:
-        raise MlflowException(
-            message=(
-                "The specified model does not contain any of the supported flavors for"
-                " deployment. The model contains the following flavors: {model_flavors}."
-                " Supported flavors: {supported_flavors}".format(
-                    model_flavors=model_config.flavors.keys(),
-                    supported_flavors=SUPPORTED_DEPLOYMENT_FLAVORS)),
-            error_code=RESOURCE_DOES_NOT_EXIST)
-
-
-def _validate_deployment_flavor(model_config, flavor):
-    """
-    Checks that the specified flavor is a supported deployment flavor
-    and is contained in the specified model. If one of these conditions
-    is not met, an exception is thrown.
-
-    :param model_config: An MLflow Model object
-    :param flavor: The deployment flavor to validate
-    """
-    if flavor not in SUPPORTED_DEPLOYMENT_FLAVORS:
-        raise MlflowException(
-            message=(
-                "The specified flavor: `{flavor_name}` is not supported for deployment."
-                " Please use one of the supported flavors: {supported_flavor_names}".format(
-                    flavor_name=flavor,
-                    supported_flavor_names=SUPPORTED_DEPLOYMENT_FLAVORS)),
-            error_code=INVALID_PARAMETER_VALUE)
-    elif flavor not in model_config.flavors:
-        raise MlflowException(
-            message=("The specified model does not contain the specified deployment flavor:"
-                     " `{flavor_name}`. Please use one of the following deployment flavors"
-                     " that the model contains: {model_flavors}".format(
-                         flavor_name=flavor, model_flavors=model_config.flavors.keys())),
-            error_code=RESOURCE_DOES_NOT_EXIST)
 
 
 def _get_mlflow_install_step(dockerfile_context_dir, mlflow_home):
@@ -136,12 +81,13 @@ def _get_mlflow_install_step(dockerfile_context_dir, mlflow_home):
         ).format(version=mlflow.version.VERSION)
 
 
-def _build_image(image_name=DEFAULT_IMAGE_NAME, mlflow_home=None, custom_setup_steps_hook=None):
+def _build_image(image_name, entrypoint, mlflow_home=None, custom_setup_steps_hook=None):
     """
     Build an MLflow Docker image that can be used to serve a
     The image is built locally and it requires Docker to run.
 
     :param image_name: Docker image name.
+    :param entry_point: String containing ENTRYPOINT directive for docker image
     :param mlflow_home: (Optional) Path to a local copy of the MLflow GitHub repository.
                         If specified, the image will install MLflow from this directory.
                         If None, it will install MLflow from pip.
@@ -156,7 +102,8 @@ def _build_image(image_name=DEFAULT_IMAGE_NAME, mlflow_home=None, custom_setup_s
         custom_setup_steps = custom_setup_steps_hook(cwd) if custom_setup_steps_hook else ""
         with open(os.path.join(cwd, "Dockerfile"), "w") as f:
             f.write(_DOCKERFILE_TEMPLATE.format(
-                install_mlflow=install_mlflow, custom_setup_steps=custom_setup_steps))
+                install_mlflow=install_mlflow, custom_setup_steps=custom_setup_steps,
+                entrypoint=entrypoint))
         _logger.info("Building docker image with name %s", image_name)
         os.system('find {cwd}/'.format(cwd=cwd))
         proc = Popen(["docker", "build", "-t", image_name, "-f", "Dockerfile", "."],
