+++ mlflow/sagemaker/__init__.py
@@ -12,11 +12,11 @@ import time
 
 import mlflow
 import mlflow.version
+from mlflow import pyfunc, mleap
 from mlflow.exceptions import MlflowException
 from mlflow.models import Model
-from mlflow.models.docker_utils import DEFAULT_IMAGE_NAME, _get_preferred_deployment_flavor,\
-    _validate_deployment_flavor
-from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
+from mlflow.models.docker_utils import DEFAULT_IMAGE_NAME
+from mlflow.protos.databricks_pb2 import RESOURCE_DOES_NOT_EXIST, INVALID_PARAMETER_VALUE
 from mlflow.tracking.artifact_utils import _download_artifact_from_uri
 from mlflow.utils import get_unique_resource_id
 from mlflow.utils.file_utils import TempDir
@@ -47,6 +47,56 @@ DEFAULT_SAGEMAKER_INSTANCE_COUNT = 1
 _logger = logging.getLogger(__name__)
 
 _full_template = "{account}.dkr.ecr.{region}.amazonaws.com/{image}:{version}"
+
+
+def _get_preferred_deployment_flavor(model_config):
+    """
+    Obtains the flavor that MLflow would prefer to use when deploying the model.
+    If the model does not contain any supported flavors for deployment, an exception
+    will be thrown.
+
+    :param model_config: An MLflow model object
+    :return: The name of the preferred deployment flavor for the specified model
+    """
+    if mleap.FLAVOR_NAME in model_config.flavors:
+        return mleap.FLAVOR_NAME
+    elif pyfunc.FLAVOR_NAME in model_config.flavors:
+        return pyfunc.FLAVOR_NAME
+    else:
+        raise MlflowException(
+            message=(
+                "The specified model does not contain any of the supported flavors for"
+                " deployment. The model contains the following flavors: {model_flavors}."
+                " Supported flavors: {supported_flavors}".format(
+                    model_flavors=model_config.flavors.keys(),
+                    supported_flavors=SUPPORTED_DEPLOYMENT_FLAVORS)),
+            error_code=RESOURCE_DOES_NOT_EXIST)
+
+
+def _validate_deployment_flavor(model_config, flavor):
+    """
+    Checks that the specified flavor is a supported deployment flavor
+    and is contained in the specified model. If one of these conditions
+    is not met, an exception is thrown.
+
+    :param model_config: An MLflow Model object
+    :param flavor: The deployment flavor to validate
+    """
+    if flavor not in SUPPORTED_DEPLOYMENT_FLAVORS:
+        raise MlflowException(
+            message=(
+                "The specified flavor: `{flavor_name}` is not supported for deployment."
+                " Please use one of the supported flavors: {supported_flavor_names}".format(
+                    flavor_name=flavor,
+                    supported_flavor_names=SUPPORTED_DEPLOYMENT_FLAVORS)),
+            error_code=INVALID_PARAMETER_VALUE)
+    elif flavor not in model_config.flavors:
+        raise MlflowException(
+            message=("The specified model does not contain the specified deployment flavor:"
+                     " `{flavor_name}`. Please use one of the following deployment flavors"
+                     " that the model contains: {model_flavors}".format(
+                        flavor_name=flavor, model_flavors=model_config.flavors.keys())),
+            error_code=RESOURCE_DOES_NOT_EXIST)
 
 
 def push_image_to_ecr(image=DEFAULT_IMAGE_NAME):
