+++ mlflow/sagemaker/cli.py
@@ -166,9 +166,14 @@ def build_and_push_container(build, push, container, mlflow_home):
     if not (build or push):
         print("skipping both build and push, have nothing to do!")
     if build:
+        sagemaker_image_entrypoint = """
+        ENTRYPOINT ["python", "-c", "import sys; from mlflow.sagemaker import container as C; \
+        C._init(sys.argv[1])"]
+        """
         mlflow.models.docker_utils._build_image(
             container,
-            mlflow_home=os.path.abspath(mlflow_home) if mlflow_home else None
+            mlflow_home=os.path.abspath(mlflow_home) if mlflow_home else None,
+            entrypoint=sagemaker_image_entrypoint,
         )
     if push:
         mlflow.sagemaker.push_image_to_ecr(container)
