+++ tests/models/test_cli.py
@@ -27,7 +27,6 @@ from mlflow.utils.file_utils import TempDir, path_to_local_file_uri
 from mlflow.utils.environment import _mlflow_conda_env
 from mlflow.utils import PYTHON_VERSION
 from tests.models import test_pyfunc
-from tests.projects.utils import tracking_uri_mock  # pylint: disable=unused-import
 from tests.helper_functions import pyfunc_build_image, pyfunc_serve_from_docker_image,\
     _evaluate_scoring_proc, get_safe_port, pyfunc_serve_and_score_model
 from mlflow.pyfunc.scoring_server import CONTENT_TYPE_JSON_SPLIT_ORIENTED,\
@@ -256,9 +255,10 @@ def test_build_docker(iris_data, sk_model):
             "and body %s" % (scoring_response.status_code, scoring_response.text)
     # Pass "python-function" as the flavor, it should work
     pyfunc_build_image(model_uri, extra_args=["-f", pyfunc.FLAVOR_NAME])
-    # Pass a non-existent flavor, expect failure
-    bad_flavor_image_name = uuid.uuid4().hex
-    res = CliRunner().invoke(build_docker, ["-m", model_uri, "-n", bad_flavor_image_name, "-f",
-                                            "some-nonexistent-flavor"])
-    assert res.exit_code != 0
-    assert "Unable to find flavor backend to serve model flavor" in str(res.exception)
+    # Pass currently-unsupported flavors (mleap, R's crate flavor), expect failure
+    for bad_flavor in ["mleap", "crate"]:
+        bad_flavor_image_name = uuid.uuid4().hex
+        res = CliRunner().invoke(build_docker, ["-m", model_uri, "-n", bad_flavor_image_name, "-f",
+                                                bad_flavor])
+        assert res.exit_code != 0
+        assert "Unable to find flavor backend to serve model flavor" in str(res.exception)
